refactor(e2e): move shared clients to function
