+++ e2e-tests/lib/addMinute.ts
@@ -1,0 +1,2 @@
+export const addMinute = (now: Date, addMinutes = 0): Date =>
+	new Date(now.getTime() + addMinutes * 60 * 1000)
