+++ e2e-tests/lib/fetchHistoricalData.ts
