+++ e2e-tests/lib/randomUsage.ts
@@ -1,0 +1,1 @@
+export const randomUsage = (): number => Math.floor(Math.random() * 1000)
